requirements.txt
Django==2.2.13
djangorestframework==3.11.0
pytz==2018.9
